+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file based on the
   All is based on docker contaienrs now
 - Support for PHP 5.3 removed
 - Elastica\Reponse::getError() now returns and array instead of a string
+- Move function \Elastica\Index\Status::getAliases() and hasAlias(...) to \Elastica\Index::getAliases()
 
 ### Bugfixes